fix(ViewportGrid):Make the set display sets work as an update as well
The set display sets hard coded the change to the viewport/display set
options if it wasn't provided, instead of doing the correct thing, which
is to pass it along empty and let the upstream merge the old values with
the new.  This allows both replace and update use cases to work.